Move get_session to dependencies alongside the other request accessors
